Add C3 enhanced firmware: apps, games, BadUSB, protocols, and more
Major feature additions to M1 firmware:

Sub-GHz: 30+ protocol decoders (CAME, Nice Flo, Keeloq, Security+ 2.0,
Linear, Holtek, Somfy, etc.), spectrum analyzer, RSSI meter, frequency
scanner, weather station decoding, extended band support (150/200/250 MHz)

RFID: 20+ protocol decoders (HID Generic, Indala, AWID, Pyramid, Paradox,
IOProx, FDX-A/B, Viking, Gallagher, etc.), clone/erase T5577, RFID fuzzer,
Manchester decoder with ASK/PSK auto-detection

NFC: Tag Info with manufacturer lookup and SAK decode, T2T page dump,
clone/emulate, NFC fuzzer, MIFARE Classic Crypto1

Infrared: Universal remote database (Samsung, LG, Sony, Bose, Denon, etc.),
learn and save signals, Flipper .ir file import

BadUSB: DuckyScript interpreter over USB HID
Bad-BT: DuckyScript over Bluetooth HID (experimental)
Games: Snake, Tetris, T-Rex Runner, Pong, Dice
External Apps: ELF loader for .m1app files from SD card
WiFi: Scan, connect, saved networks, status display
NFC/RFID Field Detector: Detect external reader fields
BT Device Manager: Scan, save, and manage BLE devices
RPC Protocol: Binary protocol for qMonstatek companion app
Dual Boot: CRC-validated bank swap with safe fallback
Settings: Persistent preferences, southpaw mode

+++ m1_csrc/m1_infrared.h
@@ -107,8 +107,8 @@ extern TIM_HandleTypeDef    Timerhdl_IrTx;
 extern TIM_HandleTypeDef    Timerhdl_IrRx;
 
 extern volatile uint8_t ir_ota_data_tx_active;
-extern uint8_t ir_ota_data_tx_len;
-extern volatile uint8_t ir_ota_data_tx_counter;
+extern uint16_t ir_ota_data_tx_len;
+extern volatile uint16_t ir_ota_data_tx_counter;
 extern uint16_t *pir_ota_data_tx_buffer;
 
 #endif /* M1_INFRARED_H_ */
